myolap-server/myolap-bg/requirements.txt
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.18.3
pandas==1.0.3
pip==20.0.2
PyMySQL==0.9.3
python-dateutil==2.8.1
pytz==2020.1
setuptools==40.8.0
six==1.14.0
Werkzeug==1.0.1
DBUtils==1.3
celery==4.3.0
redis==3.5.1
eventlet==0.25.1
#sqlalchemy==1.3.16
flask-sqlalchemy==2.4.1
gunicorn==20.0.4
gevent==20.5.0
nacos-sdk-python==0.1.1
apscheduler==3.6.3
flask_apscheduler==1.11.0
flask-jwt-extended==3.20.0
openpyxl==3.0.4

